## Description
This change improves compatibility with 32-bit architectures by
addressing integer overflow issues.

Several constants and arithmetic expressions that previously relied on
platform-dependent integer sizing have been updated to use explicit safe
types, ensuring consistent behaviour across both 32-bit and 64-bit
systems.

## Tests
- Added happy tests for the functions using `ExpectedKDFStrength`.

+++ server/ente/user.go
@@ -12,7 +12,7 @@ const (
 	SignUpOTTPurpose      = "signup"
 	LoginOTTPurpose       = "login"
 
-	ExpectedKDFStrength = 1073741824 * 4
+	ExpectedKDFStrength = int64(1073741824 * 4)
 )
 
 // User represents a user in the system
@@ -80,8 +80,8 @@ type KeyAttributes struct {
 	PublicKey                         string `json:"publicKey" binding:"required"`
 	EncryptedSecretKey                string `json:"encryptedSecretKey" binding:"required"`
 	SecretKeyDecryptionNonce          string `json:"secretKeyDecryptionNonce" binding:"required"`
-	MemLimit                          int    `json:"memLimit" binding:"required"`
-	OpsLimit                          int    `json:"opsLimit" binding:"required"`
+	MemLimit                          int64  `json:"memLimit" binding:"required"`
+	OpsLimit                          int64  `json:"opsLimit" binding:"required"`
 	MasterKeyEncryptedWithRecoveryKey string `json:"masterKeyEncryptedWithRecoveryKey"`
 	MasterKeyDecryptionNonce          string `json:"masterKeyDecryptionNonce"`
 	RecoveryKeyEncryptedWithMasterKey string `json:"recoveryKeyEncryptedWithMasterKey"`
@@ -114,8 +114,8 @@ type UpdateKeysRequest struct {
 	KEKSalt            string `json:"kekSalt" binding:"required"`
 	EncryptedKey       string `json:"encryptedKey" binding:"required"`
 	KeyDecryptionNonce string `json:"keyDecryptionNonce" binding:"required"`
-	MemLimit           int    `json:"memLimit" binding:"required"`
-	OpsLimit           int    `json:"opsLimit" binding:"required"`
+	MemLimit           int64  `json:"memLimit" binding:"required"`
+	OpsLimit           int64  `json:"opsLimit" binding:"required"`
 }
 
 func (u *UpdateKeysRequest) Validate() error {
